HackathonSingleProjects.jsx - removed style components and import
